This is ready for review
I think I am about done with this exploration.  If you can find a solution to the local install failure on the second last line that would be cool.  CI will struggle to know which installer you are making changes to unless you put each installer into a repo of its own.  I will come back if I discover how to the the return code from the powershell run.  Good luck with it and I hope this helps.